requirements/dev.txt
# Everything the developer needs in addition to the production requirements
-r prod.txt

# Management script
Flask-Script==0.6.6

# Testing
Flask-Testing==0.4
Flask-WebTest==0.0.6
WebOb==1.3.1
WebTest==2.0.10
beautifulsoup4==4.3.2
blinker==1.3
coverage==3.7.1
nose==1.3.0
twill==0.9
waitress==0.8.8
